Added main function as seprate file
